Rename jsNativeMain to nonJvmMain
