+++ compose/animation/animation-core/build.gradle
@@ -120,9 +120,9 @@ if (AndroidXComposePlugin.isMultiplatformEnabled(project)) {
                     implementation(libs.kotlinStdlib)
                 }
             }
-            jsNativeMain.dependsOn(commonMain)
-            nativeMain.dependsOn(jsNativeMain)
-            webMain.dependsOn(jsNativeMain)
+            nonJvmMain.dependsOn(commonMain)
+            nativeMain.dependsOn(nonJvmMain)
+            webMain.dependsOn(nonJvmMain)
 
             jsMain {
                 dependsOn(webMain)
@@ -135,7 +135,7 @@ if (AndroidXComposePlugin.isMultiplatformEnabled(project)) {
                 }
             }
             desktopMain.dependsOn(jbMain)
-            jsNativeMain.dependsOn(jbMain)
+            nonJvmMain.dependsOn(jbMain)
 
             wasmJsMain {
                 dependsOn(webMain)
