+++ compose/foundation/foundation-layout/build.gradle
@@ -106,11 +106,11 @@ if(AndroidXComposePlugin.isMultiplatformEnabled(project)) {
                 implementation(libs.kotlinStdlib)
             }
 
-            jsNativeMain.dependsOn(commonMain)
-            nativeMain.dependsOn(jsNativeMain)
+            nonJvmMain.dependsOn(commonMain)
+            nativeMain.dependsOn(nonJvmMain)
 
             webMain {
-                dependsOn(jsNativeMain)
+                dependsOn(nonJvmMain)
                 dependencies {
                     implementation(libs.kotlinStdlib)
                 }
