+++ compose/foundation/foundation/src/nonJvmMain/kotlin/androidx/compose/foundation/internal/JvmDefaultWithCompatibility.nonJvm.kt
