+++ compose/foundation/foundation/src/nonJvmMain/kotlin/androidx/compose/foundation/internal/PlatformOptimizedCancellationException.nonJvm.kt
