+++ compose/foundation/foundation/src/nonJvmMain/kotlin/androidx/compose/foundation/text/input/internal/CodepointHelpers.nonJvm.kt
